+++ src/DataObjects/PreviewSubscriptionPlanChangeData.php
@@ -8,15 +8,11 @@ use Subster\PhpSdk\Concerns\Data;
 use Subster\PhpSdk\Enums\PaymentStrategy;
 use Subster\PhpSdk\Enums\SubscriptionPlanChangeProrationBehavior;
 
-class ChangeSubscriptionPlanData extends Data
+class PreviewSubscriptionPlanChangeData extends Data
 {
     public function __construct(
         public readonly string $plan,
-        public readonly ?string $success_url = null,
-        public readonly ?string $cancel_url = null,
         public readonly ?SubscriptionPlanChangeProrationBehavior $proration_behavior = null,
         public readonly ?PaymentStrategy $payment_strategy = null,
-        public readonly ?string $idempotency_key = null,
-        public readonly ?string $quote = null,
     ) {}
 }