+++ src/Resources/SubscriptionsResource.php
@@ -8,14 +8,30 @@ use Saloon\Exceptions\Request\FatalRequestException;
 use Saloon\Exceptions\Request\RequestException;
 use Saloon\Http\BaseResource;
 use Subster\PhpSdk\DataObjects\ChangeSubscriptionPlanData;
+use Subster\PhpSdk\DataObjects\PreviewSubscriptionPlanChangeData;
 use Subster\PhpSdk\DataObjects\RecordSubscriptionUsageEventData;
 use Subster\PhpSdk\DataObjects\SubscriptionPlanChangeData;
+use Subster\PhpSdk\DataObjects\SubscriptionPlanChangeQuoteData;
 use Subster\PhpSdk\DataObjects\SubscriptionUsageEventData;
 use Subster\PhpSdk\Requests\ChangeSubscriptionPlanRequest;
+use Subster\PhpSdk\Requests\PreviewSubscriptionPlanChangeRequest;
 use Subster\PhpSdk\Requests\RecordSubscriptionUsageEventRequest;
 
 class SubscriptionsResource extends BaseResource
 {
+    /**
+     * @throws FatalRequestException
+     * @throws RequestException
+     */
+    public function previewPlanChange(
+        string $subscription,
+        PreviewSubscriptionPlanChangeData $data,
+    ): SubscriptionPlanChangeQuoteData {
+        return $this->connector->send(
+            new PreviewSubscriptionPlanChangeRequest($subscription, $data)
+        )->dtoOrFail();
+    }
+
     /**
      * @throws FatalRequestException
      * @throws RequestException